pipistrello: Cleanup and modernize.
